app/requirements.txt
Django==2.2
djangorestframework==3.9.4
django-cors-headers==3.0.0
django_celery_beat==1.4.0
djangorestframework-simplejwt==4.2.0
django-reversion==3.0.4
psycopg2==2.8.2
PyJWT==1.7.1
